Add async transactions - failed attempt
Creating a new instance of SQLiteAsyncConnection (member of SQLite-net-pcl package) is only saving the connectionstring.  Only when one uses "asyncConn.ExecuteAsync", a real connection is created or reused from a pool.  It looks like not every execution of 'ExecuteAsync' on the same asyncConn-instance, necessarily gets the same connection, so using this command to start/end a transaction by sending "BEGIN TRANSACTION" en "END TRANSACTION" to SQLite is not working.