Add async transactions - another failed attempt
Now using 'SQLite-net-pcl's own 'RunInTransactionAsync'.
Looks like connections are still getting mixed up, because 'asyncConn.CloseAsync()' closes the wrong connection.
If we don't close the connection, no error, but there is still a problem with isolation between read- and write transactions, and there is a lack of concurrency for read transactions..